server/status_store/app/requirements.txt
paho-mqtt
PyMySQL

